old/api/requirements.txt
fastapi
python-dotenv
requests
sqlmodel
uvicorn
ulid-py
